fix(api): keep mbti question projection scoped
